chore: move transport to its own module
