+++ lib/agent.js
@@ -6,33 +6,34 @@
 
 'use strict'
 
-var http = require('http')
-var path = require('path')
+const http = require('http')
+const path = require('path')
 
-var isError = require('core-util-is').isError
-var Filters = require('object-filter-sequence')
+const isError = require('core-util-is').isError
+const Filters = require('object-filter-sequence')
 
 const { agentActivationMethodFromStartStack } = require('./activation-method')
 const { CAPTURE_ERROR_LOG_STACK_TRACES_ALWAYS, CAPTURE_ERROR_LOG_STACK_TRACES_MESSAGES } = require('./config/schema')
-var config = require('./config/config')
-var connect = require('./middleware/connect')
+const config = require('./config/config')
+const connect = require('./middleware/connect')
 const constants = require('./constants')
-var errors = require('./errors')
+const errors = require('./errors')
 const { InflightEventSet } = require('./InflightEventSet')
-var Instrumentation = require('./instrumentation')
-var { elasticApmAwsLambda } = require('./lambda')
-var Metrics = require('./metrics')
-var parsers = require('./parsers')
-var symbols = require('./symbols')
+const Instrumentation = require('./instrumentation')
+const { elasticApmAwsLambda } = require('./lambda')
+const Metrics = require('./metrics')
+const parsers = require('./parsers')
+const symbols = require('./symbols')
 const { frameCacheStats, initStackTraceCollection } = require('./stacktraces')
 const Span = require('./instrumentation/span')
 const Transaction = require('./instrumentation/transaction')
 const { isOTelMetricsFeatSupported, createOTelMeterProvider } = require('./opentelemetry-metrics')
+const { createTransport } = require('./transport/transport')
 
-var IncomingMessage = http.IncomingMessage
-var ServerResponse = http.ServerResponse
+const IncomingMessage = http.IncomingMessage
+const ServerResponse = http.ServerResponse
 
-var version = require('../package').version
+const version = require('../package').version
 
 // ---- Agent
 
@@ -279,7 +280,7 @@ Agent.prototype.start = function (opts) {
   }
 
   initStackTraceCollection()
-  this._transport = this._conf.transport(this._conf, this)
+  this._transport = createTransport(this._conf, this)
 
   let runContextClass
   if (this._conf.opentelemetryBridgeEnabled) {
