+++ test/agent.test.js
@@ -27,7 +27,7 @@ const {
 } = require('../lib/config/schema')
 const { findObjInArray } = require('./_utils')
 const { MockAPMServer } = require('./_mock_apm_server')
-const { NoopTransport } = require('../lib/noop-transport')
+const { NoopTransport } = require('../lib/transport/noop-transport')
 var packageJson = require('../package.json')
 
 // Options to pass to `agent.start()` to turn off some default agent behavior
