+++ test/context-propagation-only.test.js
@@ -16,7 +16,7 @@ const pino = require('pino')
 const tape = require('tape')
 
 const apm = require('../')
-const { NoopTransport } = require('../lib/noop-transport')
+const { NoopTransport } = require('../lib/transport/noop-transport')
 const { MockAPMServer } = require('./_mock_apm_server')
 
 tape.test('contextPropagationOnly', function (suite) {
